deps: migrate to docusaurus 3
