--- src/theme/DocPage/Layout/index.js
@@ -1,28 +1,0 @@
-import React, { useState } from 'react';
-import { useDocsSidebar } from '@docusaurus/theme-common/internal';
-import Layout from '@theme/Layout';
-import BackToTopButton from '@theme/BackToTopButton';
-import DocPageLayoutSidebar from '@theme/DocPage/Layout/Sidebar';
-import DocPageLayoutMain from '@theme/DocPage/Layout/Main';
-import styles from './styles.module.css';
-
-export default function DocPageLayout({ children }) {
-  const sidebar = useDocsSidebar();
-  const [hiddenSidebarContainer, setHiddenSidebarContainer] = useState(false);
-
-  return (
-    <Layout wrapperClassName={styles.docsWrapper}>
-      <BackToTopButton />
-      <div className={styles.docPage}>
-          <DocPageLayoutSidebar
-            sidebar={sidebar.items}
-            hiddenSidebarContainer={hiddenSidebarContainer}
-            setHiddenSidebarContainer={setHiddenSidebarContainer}
-          />
-        <DocPageLayoutMain hiddenSidebarContainer={hiddenSidebarContainer}>
-          {children}
-        </DocPageLayoutMain>
-      </div>
-    </Layout>
-  );
-}
