+++ src/theme/DocRoot/Layout/Sidebar/ExpandButton/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import {translate} from '@docusaurus/Translate';
+import { translate } from '@docusaurus/Translate';
 import IconArrow from '@theme/Icon/Arrow';
 import styles from './styles.module.css';
 
-export default function DocPageLayoutSidebarExpandButton({toggleSidebar}) {
+export default function DocRootLayoutSidebarExpandButton({ toggleSidebar }) {
   return (
     <div
       className={styles.expandButton}
