+++ src/theme/DocRoot/Layout/Sidebar/styles.module.css
@@ -12,8 +12,6 @@
     display: block;
     width: var(--doc-sidebar-width);
     margin-top: calc(-1 * var(--ifm-navbar-height));
-    will-change: width;
-    transition: width var(--ifm-transition-fast) ease;
     clip-path: inset(0);
   }
