+++ src/theme/DocRoot/Layout/index.js
@@ -1,0 +1,27 @@
+import React, { useState } from 'react';
+import { useDocsSidebar } from '@docusaurus/theme-common/internal';
+import BackToTopButton from '@theme/BackToTopButton';
+import DocRootLayoutSidebar from '@theme/DocRoot/Layout/Sidebar';
+import DocRootLayoutMain from '@theme/DocRoot/Layout/Main';
+import styles from './styles.module.css';
+
+export default function DocRootLayout({ children }) {
+  const sidebar = useDocsSidebar();
+  const [hiddenSidebarContainer, setHiddenSidebarContainer] = useState(false);
+
+  return (
+    <div className={styles.docsWrapper}>
+      <BackToTopButton />
+      <div className={styles.docRoot}>
+        <DocRootLayoutSidebar
+          sidebar={sidebar.items}
+          hiddenSidebarContainer={hiddenSidebarContainer}
+          setHiddenSidebarContainer={setHiddenSidebarContainer}
+        />
+        <DocRootLayoutMain hiddenSidebarContainer={hiddenSidebarContainer}>
+          {children}
+        </DocRootLayoutMain>
+      </div>
+    </div>
+  );
+}
