+++ src/theme/DocRoot/Layout/styles.module.css
@@ -1,4 +1,4 @@
-.docPage {
+.docRoot {
   display: flex;
   flex: 1 0;
   justify-content: center;
@@ -7,8 +7,4 @@
 .docsWrapper {
   display: flex;
   flex: 1 0 auto;
-}
-
-html:not([data-theme]) .themedComponent--light {
-  display: initial;
 }
